moved about page and home page calls to db folder
