+++ lib/db/cms/home-page.ts
@@ -1,5 +1,24 @@
-const getHomePage = async () => {
-    //Move Supa base about page here
+import { supabase } from '@/lib/supabase'
+
+const getHomePageData = async () => {
+    try {
+        const { data, error } = await supabase
+            .from('home_page')
+            .select('heroBanner, heroImgAltText, heroImgSrc')
+
+        if (error) throw error
+
+        return {
+            heroBanner: data[0].heroBanner,
+            heroImg: {
+                altText: data[0].heroImgAltText,
+                src: data[0].heroImgSrc
+            }
+        }
+    } catch (error) {
+        console.error(error)
+        return {}
+    }
 }
 
-export { getHomePage }
+export { getHomePageData }
