+++ views/404View/index.tsx
@@ -1,0 +1,5 @@
+const ErrorPage = ({ code }: { code: string | number }) => {
+    return <div>`A ${code} error has occured.`</div>
+}
+
+export default ErrorPage
